requirements.txt
streamlit==1.31.0
langchain==0.1.12
langchain-openai==0.0.5
langchain-core==0.1.32
openai==1.13.3
python-dotenv==1.0.0
PyPDF2==3.0.1
playwright==1.41.2
requests==2.31.0
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
pre-commit==3.5.0
typing-extensions==4.8.0
python-multipart==0.0.6
tqdm==4.66.1